requirements.txt
# 核心依赖
PyYAML>=6.0
requests>=2.28.0
python-dotenv>=0.19.0

# 文件处理（可选）
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.7.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# 文本处理
nltk>=3.8
jieba>=0.42.1
regex>=2022.7.0

# 性能优化
aiohttp>=3.8.0
asyncio>=3.4.3

# 缓存
redis>=4.3.0
diskcache>=5.4.0

# 日志
loguru>=0.6.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.20.0

# 开发工具
black>=22.0.0
flake8>=5.0.0
mypy>=0.971